requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.6
pyyaml==6.0.2
apscheduler==3.10.4
python-multipart==0.0.20
exa-py
beautifulsoup4
lxml
chromadb
anthropic>=0.49.0
tiktoken>=0.7.0
pytest==8.3.4
